requirements.txt
Django==4.0.4
python-dotenv==0.20.0
gino==1.0.1
aiogram==2.20
loader==2017.9.11
SQLAlchemy==1.3.24
asyncpg==0.25.0
psycopg2==2.9.3